mostly meds and chemicals recipes
slight xhange to the enviro suit so it allows mutation attacks. i think it would be aproproate for a pausi-intelligent slime suit.
a shift to the multi tool to allow it to dig and to prepare it potentially for the graft bionics.
we now have nati fungal drugs and fertilizer from crap.